made sure that there is no float vs double vs decimal confusion for lat/lon

+++ server/backend/migration/static_data/source-schema.sql
@@ -125,8 +125,8 @@ create table if not exists feedback
     reply_to    text charset utf8                                      null,
     feedback    text charset utf8                                      not null,
     image_count int                                                    null,
-    latitude    decimal(11, 8)                                         null,
-    longitude   decimal(11, 8)                                         null,
+    latitude    double                                                 null,
+    longitude   double                                                 null,
     timestamp   datetime /* mariadb-5.3 */ default current_timestamp() null
 ) auto_increment = 293;
